requirements.txt
asgiref==3.3.3
beautifulsoup4==4.9.3
boto3==1.17.62
botocore==1.20.62
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-embed-video==1.4.0
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
wincertstore==0.2
